[APP-2865] Extract analytics logic from ApkfyViewModel
